+++ avantis_trader_sdk/__init__.py
@@ -1,5 +1,6 @@
 from .client import TraderClient
 from .feed.feed_client import FeedClient
 from .signers.base import BaseSigner
+from .types import PriceSourcing
 
-__version__ = "0.8.13"
+__version__ = "0.8.14"